Add filters::builtin mod to make hiding impl details easier
